Move to new account/user style
Xednese will now move into a style that separate account-based actions from
user-based to make it more obvious. This means that instead of creating an
Esendex instance with an account reference you will have to call #account
instead.

Hopefully this will make it easier to understand, when say, getting received
messages as you will know whether you are asking in the context of an account or
across all a user's accounts.